refactor: use-lock-scroll to scroll-lock
